# docs/swagger.yaml
definitions:
  errcode.Error:
    type: object
  model.ListResponse-model_UserBloodPressure:
    properties:
      count:
        type: integer
      data:
        items:
          $ref: '#/definitions/model.UserBloodPressure'
        type: array
    type: object
  model.ListResponse-model_WeightRecordVO:
    properties:
      count:
        type: integer
      data:
        items:
          $ref: '#/definitions/model.WeightRecordVO'
        type: array
    type: object
...
  model.SuccessReponseSwagger:
    properties:
      success:
        type: boolean
    type: object
  model.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        type: string
      id:
        type: string
      lastName:
        type: string
      password:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    required:
    - firstName
    - gender
    - lastName
    - password
    - username
    type: object
  model.UserBloodPressure:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      diastolicBloodPressure:
        type: integer
      id:
        type: string
      logDatetime:
        type: string
      pulse:
        type: integer
      systolicBloodPressure:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userId:
        type: string
    type: object
...
  model.UserTemperature:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      notes:
        type: string
      recordDate:
        type: string
      temperature:
        type: number
      unit:
        description: e.g. "C" or "F"
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userId:
        type: string
    type: object
...
  model.WeightRecord:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      recordDate:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
      weight:
        maximum: 500
        type: number
    required:
    - weight
    type: object
  model.WeightRecordVO:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: string
      recordDate:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/model.User'
      userId:
        type: string
      weight:
        maximum: 500
        type: number
    required:
    - weight
    type: object
info:
  contact: {}
  description: health master open api
  termsOfService: http://swagger.io/terms/
  title: health master api
  version: "1.0"
paths:
...
  /api/v1/user-temperature:
    post:
      consumes:
      - application/json
      description: create a new record for user temperature
      parameters:
      - description: create a user temperature record
        in: body
        name: user_temperature
        required: true
        schema:
          $ref: '#/definitions/model.UserTemperature'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/model.UserTemperature'
        "400":
          description: bad request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      security:
      - BearerAuth: []
      summary: create a user temperature record
      tags:
      - user_temperature
  /api/v1/user-temperature/{id}:
    get:
      consumes:
      - application/json
      description: get a user temperature record by id
      parameters:
      - description: user temperature id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/model.UserTemperature'
        "400":
          description: bad request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      security:
      - BearerAuth: []
      summary: get a user temperature record
      tags:
      - user_temperature
  /api/v1/user-temperatures:
    get:
      consumes:
      - application/json
      description: list all user temperature records for a user
      parameters:
      - description: user id
        in: query
        name: userId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            items:
              $ref: '#/definitions/model.UserTemperature'
            type: array
        "400":
          description: bad request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      security:
      - BearerAuth: []
      summary: list all user temperature records
      tags:
      - user_temperature
...
  /api/v1/user_blood_pressure:
    post:
      consumes:
      - application/json
      description: create a user blood pressure
      parameters:
      - description: create a user blood pressure
        in: body
        name: user_blood_pressure
        required: true
        schema:
          $ref: '#/definitions/model.UserBloodPressure'
      produces:
      - application/json
      responses:
        "200":
          description: sucess
          schema:
            $ref: '#/definitions/model.UserBloodPressure'
        "400":
          description: bad request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      security:
      - BearerAuth: []
      summary: create a user blood pressure
      tags:
      - user_blood_pressure
  /api/v1/user_blood_pressure/{id}:
    post:
      consumes:
      - application/json
      description: delete a user blood pressure record
      parameters:
      - description: user blood pressure id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: sucess
          schema:
            items:
              $ref: '#/definitions/model.SuccessReponseSwagger'
            type: array
        "400":
          description: bad request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      security:
      - BearerAuth: []
      summary: delete a user blood pressure record
      tags:
      - user_blood_pressure
  /api/v1/user_blood_pressure/{id}/export:
    get:
      consumes:
      - application/json
      description: download user blood pressure excel
      parameters:
      - description: user blood pressure id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: excel file
          schema:
            type: file
        "400":
          description: bad request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      security:
      - BearerAuth: []
      summary: download user blood pressure excel
      tags:
      - user_blood_pressure
  /api/v1/user_blood_pressures:
    post:
      consumes:
      - application/json
      description: list all user blood pressure for single user
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: sucess
          schema:
            $ref: '#/definitions/model.ListResponse-model_UserBloodPressure'
        "400":
          description: bad request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      security:
      - BearerAuth: []
      summary: list all user blood pressure for single user
      tags:
      - user_blood_pressure
  /api/v1/weight-record:
    post:
      consumes:
      - application/json
      description: create a new weight record
      parameters:
      - description: create a weight record
        in: body
        name: weight_record
        required: true
        schema:
          $ref: '#/definitions/model.WeightRecord'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/model.WeightRecord'
        "400":
          description: bad request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      security:
      - BearerAuth: []
      summary: create a weight record
      tags:
      - weight_record
  /api/v1/weight-record/{id}:
    delete:
      consumes:
      - application/json
      description: delete a weight record by id
      parameters:
      - description: weight record id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            additionalProperties: true
            type: object
        "400":
          description: bad request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      security:
      - BearerAuth: []
      summary: delete a weight record
      tags:
      - weight_record
  /api/v1/weight-records:
    get:
      consumes:
      - application/json
      description: list weight records with pagination
      parameters:
      - description: user id
        in: query
        name: userId
        type: string
      - description: access token
        in: query
        name: accessToken
        type: string
      - description: Authorization token
        in: header
        name: Authorization
        type: string
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/model.ListResponse-model_WeightRecordVO'
        "400":
          description: bad request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/errcode.Error'
      security:
      - BearerAuth: []
      summary: list weight records
      tags:
      - weight_record
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"